Add additional edge case tests for Contact and GradeAnalyzer

+++ module2/GradeAnalyzer.java
@@ -75,15 +75,16 @@ public class GradeAnalyzer {
 
     // Returns the average of a list of scores, or 0.0 if the list is empty
     public static double calculateAverage(ArrayList<Integer> scores) {
-        // your code here
         if (scores.isEmpty()) {
             return 0.0;
         }
-        double total = 0.0;
+
+        int total = 0;
         for (int score : scores) {
             total += score;
         }
-        return total / scores.size();
+
+        return (double) total / scores.size();
     }
 
     // Writes and prints the report
